requirements.txt
tweepy
nltk
python-dotenv
"tenacity>=8.2.0" 
